requirements.txt
# Core Framework
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# LangGraph & LangChain
langgraph>=0.0.26
langchain-core>=0.1.23

# HTTP Client
httpx>=0.26.0,<1.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1  # Pin to compatible version for passlib
python-multipart>=0.0.6

# Environment
python-dotenv>=1.0.0

# Async support (for compatibility)
anyio>=4.0.0
